#1. Convert numeric type from double to float #2. Refactor configuration file and command line parameter #3. use SIMD for backward pass in output layer

+++ RNNSharp/DataSet.cs
@@ -82,7 +82,7 @@ namespace RNNSharp
 
             for (int i = 0; i < TagSize; i++)
             {
-                double sum = 0;
+                float sum = 0;
                 for (int j = 0; j < TagSize; j++)
                 {
                     sum += CRFLabelBigramTransition[i][j];
